chore: move caculator into app
